+++ extensions/GUI/CCControlExtension/CCControlSwitch.cpp
@@ -174,7 +174,7 @@ void CCControlSwitchSprite::draw()
     long offset = 0;
     setGLBufferData(&m_sQuad, 4 * kQuadSize, 0);
 #else
-    long offset = (long)&m_sQuad;
+    uintptr_t offset = (uintptr_t)&m_sQuad;
 #endif // EMSCRIPTEN
 
     // vertex
